cambiar el tipo de dato de 'edad' en la estructura 'jugador_t' de short a int

+++ 03_u_evaluables/Stefan_Trifan_Practica1_IPR1/5_partidos_pinpon/main.c
@@ -23,7 +23,7 @@ typedef struct jugador_t
 {
     char nombre[TAM_STR];
     char apellidos[TAM_STR];
-    short edad;
+    int edad;
 }jugador_t;
 
 typedef struct fechas
